refactoring due to inputs of ood
