--- controllers/note-app-controller.js
@@ -1,14 +1,0 @@
-import path from "path";
-
-export class NoteAppController {
-  constructor() {
-    const rootPath = path.resolve();
-    this.noteRootPath = path.join(rootPath, "public/html/");
-  }
-
-  getIndexDocument = async (req, res) => {
-    res.sendFile(`${this.noteRootPath}index.html`);
-  };
-}
-
-export const noteAppController = new NoteAppController();
