+++ package.json
@@ -34,6 +34,7 @@
     "moment": "^2.29.4",
     "mongoose": "^7.1.0",
     "path": "0.12.7",
+    "http-errors": "^2.0.0",
     "socket.io": "^4.6.1"
   }
 }
