+++ public/index.html
@@ -8,14 +8,8 @@
       const socket = io();
     </script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.29.4/moment.min.js"></script>
-    <script
-      src="../scripts/controllers/app-controller.js"
-      type="module"
-    ></script>
-    <script
-      src="../scripts/controllers/notes-controller.js"
-      type="module"
-    ></script>
+    <script src="../js/controllers/app-controller.js" type="module"></script>
+    <script src="../js/controllers/notes-controller.js" type="module"></script>
     <link
       href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;700&display=swap"
       rel="stylesheet"
