+++ public/js/services/note.js
