+++ public/styles/main.css
@@ -51,6 +51,7 @@ footer {
 dialog {
   color: var(--font-color);
   background-image: var(--bg-image);
+  width: 30%;
 }
 
 dialog::backdrop {
